+++ backend/api/fundus_image.py
