+++ backend/app.py
@@ -2,7 +2,7 @@ from flask import Flask, jsonify
 from flask_cors import CORS
 from api.auth import auth_bp
 from routes import bp  # Import the routes Blueprint
-from api.Fundus_image import image_bp
+from backend.api.fundus_image import image_bp
 
 app = Flask(__name__)
 CORS(app)  # Enable CORS for all routes
